move confirmation dialogs
