+++ apps/studio/components/interfaces/Account/TOTPFactors/DeleteFactorModal.tsx
@@ -6,7 +6,7 @@ import {
   Modal,
 } from 'ui'
 
-import ConfirmationModal from 'components/ui/ConfirmationModal'
+import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 import { useMfaUnenrollMutation } from 'data/profile/mfa-unenroll-mutation'
 import { useStore } from 'hooks'
