+++ apps/studio/components/interfaces/Database/Triggers/CreateTrigger.tsx
@@ -7,7 +7,7 @@ import SVG from 'react-inlinesvg'
 
 import { Dictionary } from 'types'
 import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectContext'
-import ConfirmationModal from 'components/ui/ConfirmationModal'
+import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 import FormEmptyBox from 'components/ui/FormBoxEmpty'
 import NoTableState from 'components/ui/States/NoTableState'
 import {
