+++ apps/studio/components/interfaces/Settings/API/JWTSettings.tsx
@@ -31,7 +31,7 @@ import {
 } from 'ui'
 
 import { useParams } from 'common/hooks'
-import ConfirmationModal from 'components/ui/ConfirmationModal'
+import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 import Panel from 'components/ui/Panel'
 import { useJwtSecretUpdateMutation } from 'data/config/jwt-secret-update-mutation'
 import { useJwtSecretUpdatingStatusQuery } from 'data/config/jwt-secret-updating-status-query'
