+++ apps/studio/components/interfaces/Settings/Logs/Logs.SavedQueriesItem.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react'
 import toast from 'react-hot-toast'
 
 import Table from 'components/to-be-cleaned/Table'
-import ConfirmationModal from 'components/ui/ConfirmationModal'
+import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 import { useContentDeleteMutation } from 'data/content/content-delete-mutation'
 import { useContentUpsertMutation } from 'data/content/content-upsert-mutation'
 import {
