+++ apps/studio/components/interfaces/TableGridEditor/DeleteConfirmationDialogs.tsx
@@ -5,7 +5,7 @@ import { Alert, Button, Checkbox, IconExternalLink, Modal } from 'ui'
 import { SupaRow } from 'components/grid'
 import { formatFilterURLParams } from 'components/grid/SupabaseGrid.utils'
 import { useProjectContext } from 'components/layouts/ProjectLayout/ProjectContext'
-import ConfirmationModal from 'components/ui/ConfirmationModal'
+import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 import { useDatabaseColumnDeleteMutation } from 'data/database-columns/database-column-delete-mutation'
 import { useTableRowDeleteAllMutation } from 'data/table-rows/table-row-delete-all-mutation'
 import { useTableRowDeleteMutation } from 'data/table-rows/table-row-delete-mutation'
