+++ apps/studio/components/to-be-cleaned/Storage/EmptyBucketModal.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'common'
 import toast from 'react-hot-toast'
 
-import ConfirmationModal from 'components/ui/ConfirmationModal'
+import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 import { useBucketEmptyMutation } from 'data/storage/bucket-empty-mutation'
 import { Bucket } from 'data/storage/buckets-query'
 import { useStorageStore } from 'localStores/storageExplorer/StorageExplorerStore'
